requirements.txt
flask
flask-cors
faiss-cpu
sentence-transformers
transformers
torch
